[LMVP-78] Handled permissions to add and remove roles

+++ packages/leemons-plugin-media-library/src/services/permissions/remove.js
@@ -1,22 +1,35 @@
 const { permissions: table } = require('../tables');
+const canUnassignRole = require('./helpers/canUnassignRole');
+const getAssignerAndAssigneeRoles = require('./helpers/getAssignerAndAssigneeRoles');
 // const isAssetOwner = require('./helpers/isAssetOwner');
 
-module.exports = async function remove(asset, { userSession, transacting } = {}) {
-  const userAgent =
-    userSession.userAgents && userSession.userAgents.length ? userSession.userAgents[0].id : null;
+module.exports = async function remove(asset, assigneeAgent, { userSession, transacting } = {}) {
   try {
-    // if (await isAssetOwner(asset, { userSession, transacting })) {
-    //   throw new Error("You can't remove permissions from the owner");
-    // }
+    // EN: Get assigner and assignee roles
+    // ES: Obtener los roles del asignador y del asignado
+    const { assignerRole, assigneeRole } = await getAssignerAndAssigneeRoles(
+      asset,
+      userSession,
+      assigneeAgent,
+      { transacting }
+    );
 
+    // EN: Check if assigner can remove role from assignee
+    // ES: Comprobar si el asignador puede eliminar el rol del asignado
+    if (!canUnassignRole(assignerRole, assigneeRole, null)) {
+      throw new Error("You don't have permission to remove this role");
+    }
+
+    // EN: Remove role
+    // ES: Eliminar rol
     return await table.deleteMany(
       {
         asset,
-        userAgent,
+        userAgent: assigneeAgent,
       },
       { transacting }
     );
   } catch (e) {
-    throw new Error(`Failed to delete permissions: ${e.message}`);
+    throw new Error(`Failed to delete role: ${e.message}`);
   }
 };
